+++ LHCb/Det/OTDet/OTDet/RtRelation.h
@@ -11,13 +11,13 @@ namespace OTDet
   struct ValueWithError
   {
     ValueWithError() : val(0), err(0) {}
-    ValueWithError(float v, float e) : val(v), err(e) {}
-    float val ;
-    float err ;
+    ValueWithError(double v, double e) : val(v), err(e) {}
+    double val ;
+    double err ;
   } ;
 
-  inline ValueWithError interpolate( const float wlhs, const ValueWithError lhs, 
-				     const float wrhs, const ValueWithError rhs) 
+  inline ValueWithError interpolate( const double wlhs, const ValueWithError lhs, 
+				     const double wrhs, const ValueWithError rhs) 
   { 
     return ValueWithError( wlhs*lhs.val + wrhs*rhs.val, wlhs*lhs.err + wrhs*rhs.err ) ;
   }
@@ -31,35 +31,35 @@ namespace OTDet
     /// default constructor. does nothing sensible.
     RtRelation() {}
     /// constructor from a polynomial t(r). this is what you should use.
-    RtRelation(float rmax, const std::vector<double>& tcoeff, const std::vector<double>& terrcoeff, size_t nbins=25) ;
+    RtRelation(double rmax, const std::vector<double>& tcoeff, const std::vector<double>& terrcoeff, size_t nbins=25) ;
     ///  constructor from a polynomial t(r) and a fixed error in r. this is only for convenience with MC.
-    RtRelation(float rmax, const std::vector<double>& tcoeff, float sigmaR, size_t nbins=25) ;
+    RtRelation(double rmax, const std::vector<double>& tcoeff, double sigmaR, size_t nbins=25) ;
     /// constructor from a table o bins r[t]. decapricated.
-    /// RtRelation(float tmin,float tmax, const std::vector<float>& radius = boost::assign::list_of(0.0)(2.5) ) ;
+    /// RtRelation(double tmin,double tmax, const std::vector<double>& radius = boost::assign::list_of(0.0)(2.5) ) ;
     /// return radius for given drifttime
-    RadiusWithError radiusWithError(float time) const ;
+    RadiusWithError radiusWithError(double time) const ;
     /// return radius for given drifttime
-    float radius(float time) const ;
+    double radius(double time) const ;
     /// return radius for given drifttime. extrapolate into 'unphysical' region.
-    float extrapolatedradius( float time ) const ;
+    double extrapolatedradius( double time ) const ;
     /// return derivative to time
-    float drdt(float time) const ;
+    double drdt(double time) const ;
     /// return derivative to radius
-    float dtdr(float radius) const ;
+    double dtdr(double radius) const ;
     /// return drifttime for given radius
-    float drifttime( float radius) const ;
+    double drifttime( double radius) const ;
     /// return sigma(drifttime) for given radius
-    float drifttimeError( float radius) const ;
+    double drifttimeError( double radius) const ;
     /// return drifttime with error
-    DriftTimeWithError drifttimeWithError( float radius ) const ;
+    DriftTimeWithError drifttimeWithError( double radius ) const ;
     /// return average driftvelocity
-    float drdt() const { return m_rmax/(m_tmax-m_tmin) ; }
+    double drdt() const { return m_rmax/(m_tmax-m_tmin) ; }
     /// return maximum drift time
-    float tmax() const { return m_tmax ; }
+    double tmax() const { return m_tmax ; }
     /// return minimum drift time
-    float tmin() const { return m_tmin ; }
+    double tmin() const { return m_tmin ; }
     /// return maximum drift radius
-    float rmax() const { return m_rmax ; }
+    double rmax() const { return m_rmax ; }
     /// return coefficients of polynimial t(r)
     const std::vector<double>& tcoeff() const { return m_tcoeff ; }
     /// return coefficients of polynimial sigma_t(r)
@@ -71,75 +71,75 @@ namespace OTDet
       return gsl_poly_eval(&(coeff.front()),coeff.size(),x) ; }
     void initRTable(size_t ntbins) ;
   private:
-    float m_rmax ;
+    double m_rmax ;
     std::vector<double> m_tcoeff ;    /// coefficients of polynomial t(r)
     std::vector<double> m_terrcoeff ; /// coefficients of polynomial sigma_t(r)
-    //float m_rmin ;
-    float m_tmin ;
-    float m_tmax ;
-    float m_dt ;
+    //double m_rmin ;
+    double m_tmin ;
+    double m_tmax ;
+    double m_dt ;
     std::vector<RadiusWithError> m_rtable ; /// table which maps to a radius with error
   } ;
 
-  inline RadiusWithError RtRelation::radiusWithError( float time ) const 
+  inline RadiusWithError RtRelation::radiusWithError( double time ) const 
   {
-    float        fbin = (time-m_tmin)/m_dt ;
+    double        fbin = (time-m_tmin)/m_dt ;
     unsigned int ibin = int(fbin) ;
-    float        w    = fbin - ibin ;
+    double        w    = fbin - ibin ;
     return time <= m_tmin ? m_rtable.front() : 
       ( time >= m_tmax ? m_rtable.back() : 
 	interpolate( (1-w), m_rtable[ibin], w, m_rtable[ibin+1] ) ) ;
   }
   
-  inline float RtRelation::radius( float time ) const 
+  inline double RtRelation::radius( double time ) const 
   {
-    float        fbin = (time-m_tmin)/m_dt ;
+    double        fbin = (time-m_tmin)/m_dt ;
     unsigned int ibin = int(fbin) ;
-    float        w    = fbin - ibin ;
+    double        w    = fbin - ibin ;
     return time <= m_tmin ? m_rtable.front().val : 
       ( time >= m_tmax ? m_rtable.back().val : 
 	(1-w)*m_rtable[ibin].val + w*m_rtable[ibin+1].val ) ;
   }
   
-  inline float RtRelation::extrapolatedradius( float time ) const 
+  inline double RtRelation::extrapolatedradius( double time ) const 
   {
     // horribly expensive
-    float fbin = (time-m_tmin)/m_dt ;
+    double fbin = (time-m_tmin)/m_dt ;
     int ibin = int(fbin) ;
     int maxbin = m_rtable.size()-2 ;
     ibin = ibin<=0 ? 0 : (ibin >= maxbin ? maxbin: ibin) ;
-    float w    = fbin - ibin ;
+    double w    = fbin - ibin ;
     return (1-w)*m_rtable[ibin].val + w*m_rtable[ibin+1].val ;
   }
   
-  inline float RtRelation::drdt( float time ) const 
+  inline double RtRelation::drdt( double time ) const 
   {
-    float fbin = time/m_dt ;
+    double fbin = time/m_dt ;
     unsigned int ibin = std::min((unsigned int)(fbin),(unsigned int)m_rtable.size()-2) ;
     return (m_rtable[ibin+1].val - m_rtable[ibin].val)/m_dt ;
   }
   
-  inline float RtRelation::drifttime( float r ) const 
+  inline double RtRelation::drifttime( double r ) const 
   {
-    return float(polyeval(m_tcoeff,r )) ;
+    return polyeval(m_tcoeff,r ) ;
   }
 
-  inline float RtRelation::drifttimeError( float r ) const 
+  inline double RtRelation::drifttimeError( double r ) const 
   {
-    return float(polyeval(m_terrcoeff,r)) ;
+    return polyeval(m_terrcoeff,r) ;
   }
 
-  inline DriftTimeWithError RtRelation::drifttimeWithError( float r ) const 
+  inline DriftTimeWithError RtRelation::drifttimeWithError( double r ) const 
   {
     return DriftTimeWithError(drifttime(r),drifttimeError(r)) ;
   }
 
-  inline float RtRelation::dtdr( float r ) const 
+  inline double RtRelation::dtdr( double r ) const 
   {
     // we probably want to cache these ...
     std::vector<double> derivcoeff(m_tcoeff.size()-1) ;
     for( size_t i=0; i<derivcoeff.size(); ++i) derivcoeff[i] = m_tcoeff[i+1] * (i+1) ;
-    return float(polyeval(derivcoeff,r)) ;
+    return polyeval(derivcoeff,r) ;
   }
   
 }
